Widen the quantized square sums and document what the metadata describes
The byte accumulators inside quantize() are local counters, not stored fields: the
metadata slots remain four FP32 values and the blob layout is unchanged. But the
squares counter was uint32_t, and each byte contributes up to 255^2, so the total
passes UINT32_MAX just above dimension 66051. A wrapped counter feeds the double
expansion that fills sum_squares, so the stored norm and every L2 distance built on
it would be wrong rather than merely imprecise. It also fits the documented 2^16
limit by under 1%. The four chains and the combined total are now uint64_t, in the
product code and in both test mirrors of the quantizer.

The class documentation still defined x_sum and x_sum_squares as sums over the input
values, which is what this branch changed. That prose is where the defect came from:
it wrote the asymmetric L2 identity over x while the kernel's inner product term is
over the reconstruction. Both are now written in terms of x_r, and the symmetric
inner product note records that recovering the quantized sum from the stored sum is
exact rather than approximate, which is what makes a metadata-only fix sufficient.

Tests:

The regression tolerance was 1e-3 * max(1.0, expected), which collapses to a flat
1e-3 because the distances asserted run from 7.7e-7 to 2.2e-3. At dimension 4 that
is over a thousand times the value being pinned, so a kernel returning zero would
pass, and the old metadata cleared the threshold by only 1.5x. The error floor is
FP32 cancellation in sum_sq_x + sum_sq_y - 2*IP, which scales with the norms and not
with the distance, so the bound is now 8 * FLT_EPSILON * (norm_x + norm_y).

EXPECT_GE(got, 0.0f) asserted a property this change does not provide. The residual
cancellation is about 1e-7 of the norm, and two identical blobs have a true distance
of zero, so self distance lands just below it. The bound is now the noise floor
rather than zero, and the limitation is stated where it is asserted.

Added SQ8_FP32_L2_matches_reconstruction_against_float_query. The asymmetric path is
how the defect stayed hidden: its tests run dimensions 1, 5, 7 and 15 against an
absolute tolerance of 0.01, and 0.4% of the norm at dimension 5 is about 0.007, which
fits underneath. The new test runs near duplicates out to dimension 512 against a
double reference, on both the scalar and dispatched kernels.

+++ tests/unit/unit_test_utils.h
@@ -253,8 +253,11 @@ inline void ComputeSQ8Quantization(const float *original_blob, size_t dim, uint8
     // Quantize, accumulating the bytes as exact integers, then derive the sums over the
     // reconstruction min + delta * a[i], which is what the kernel algebra is written in terms of.
     // See QuantPreprocessor::quantize.
+    // 64-bit for the squares: 255^2 * dim passes UINT32_MAX just above dim 66051, and a
+    // wrapped counter would corrupt the stored norm rather than round it. Mirrors
+    // QuantPreprocessor::quantize.
     uint32_t q_sum = 0;
-    uint32_t q_sum_squares = 0;
+    uint64_t q_sum_squares = 0;
     for (size_t i = 0; i < dim; i++) {
         float normalized = (original_blob[i] - min_val) / delta;
         const uint32_t q = static_cast<uint8_t>(std::round(normalized));
